Introduce FunctionSets, BatchManager, and BatchedPartitionedKeymap
FunctionSets manage batches of functions. They only store locally the
functions that the process is involved with.

For example, with 4 functions and 2 batches across 2 processes,
each process will be involved in one batch with 2 functions each.

With 4 functions in 4 batches across 2 processes, each process
will be involved with two batches.

With 4 functions in 1 batch (similar to non-batched) each process is
involved with all 4 functions.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>