Make fcube_for_mul GPU-aware
We cannot fully parallelize it but we can parallelize at least over K.
There probably are some bad memory accesses but that remains to be seen.
The compute_scaling function has loop-carried dependencies so we
cannot parallelize that. Inlining it into fcube_for_mul helps
parallelize the mu-loop. Otherwise we would need more shared memory for
the p array.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>